Add `npm run-script export-dev`
Compiles the app, then exports the static
result into a "dist_dev" directory.

This is a "dev" rather than production build,
since it's not minified and contains source maps.

Note that this doesn't actually work at the moment,
because index.ts's "MIME type ('video/mp2t') is not executable."
In other words, we'll need to change the file extension to
.js?