Add public Firebase upload capability.
Note that this is completely unauthenticated and
public. It does not make any attempt to keep a
user's data safe, and does not download data back
from the storage bucket.

This should be treated as very much "in development"
and not ready for users yet.

In order to enable this, one must first set up a
Firebase bucket, then add the necessary configuration
(API keys, etc) into firebase.ts, and then launch
Llaminator with the query parameter "firebase-sync=true".

Fixes #102.